Update patch set 2
Patch Set 2:

@thiemo: Well, the difference is that you were going for a "fresh start" approach by ditching the old code. I prefer the opposite: keep as much of the structure of the code intact as possible, so the diff shows just the intended change in behaviour (in this case, register callbacks, not class names). You are right though that I neglected to update some of the documentation, which is bad. Leaving the unused code intact however was intentional: we can remove it before or after this patch, but I'm not sure yet that we really don't need it.

To me, "minimal change" is the more obvious approach, and also much easier to review.

But I don't think arguing about which approach is The Right One is worth our time. It is however worth our time to realize that there are different approaches, and to understand their implications, and see why different people prefer one or the other. Understanding which approach is useful in which situation (and when working with which people), and also knowing their downsides, is quite valuable.

I wrote this patch mainly to demonstrate how I would go about refactoring this. I'll not amend it for now, since I don't see the point of working on competing patches.

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0