Update patch set 11
Patch Set 11:

Seems we have a definition for an acceptable amount of constructor parameters now: Patches that introduce the 11th constructor parameter get merged (I96ec3c2), patches with 13 parameters are bikeshedded and then ignored, rot and die.

I really wanted to continue working on this, as I explained in the commit message. But when not even this first simple step is merged, I really do not see why I should waste more of my time with such implementation details. People are obviously ok with how bad the code is. Why waste my time improving our code base when all I get for that effort is an "it's still not good enough"? Surprise. Our code will *never* be "good enough". That's why we do refactoring.

This patch here is quite a good example for the level of frustration I feel. I6871b19 is an other example. The -1 is misused for "you have to fix all the stuff first before you are allowed to even go the first step". I strongly believe this is not how "agile" should work.

Patch-set: 11
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0