Update patch set 2
Patch Set 2:

I think the downside of the approach in this patch is very visible now, simply counting the number of comments. "Minimized" refactoring like this that tries to keep structures "intact" that look similar but aren't tricks reviewers into missing so many things. Code and especially documentation degenerates over time.

Which is what I wanted to avoid.

Which is why I did not wanted to rework my patch into what this now became.

As I just explained in I1aa5890 (an unrelated patch that run into the same dead end) I refuse to waste time on building chains of patches when the first step of the chain is blocked with a -1 for no reason I can fix in that patch. I really do not see why I have to "prove" my intention beyond what I wrote and what everybody can see in I1aa5890: it removes unused code and a constructor we know we can not use any more. That's all.

Patch-set: 2