Remove unused public, trivial getters
All these methods are not even tested. I understand the original intention:
All these methods are pairs of setters and getters. The setters are
required in tests. However, I do not think it's worth the overhead.

This is the regex I used to find these (one line, wrapped for readability):
grep -rhioP --include=*.php '(?<=\bpublic function )[a-z]+(?=\()'
 client/includes/ lib/includes/ purtle/src/ repo/includes/ view/src/|
 awk '{print "echo -n "$0":\ &&grep -r --include=*.php "$0" .|
 grep -v public\ function|wc -l"}'|bash|grep :\ 0

Change-Id: I649ff3a68de95c0feaa60ebbc5fa0f588d8257f3