Fix fatal error in OutputPageBeforeHTMLHookHandler
Check that we have valid EntityId and EntityRevision objects
before trying to load an Entity from EntityRevision.

There are cases (e.g. an admin viewing a revision of a deleted item)
where the EntityId is not available or an EntityRevision is not found.

(my patch on master was not a clean cherry pick so I have
made this patch against the branch)

Bug: T132645
Change-Id: I513c64648f46600e6f0b1a2e565f47775853b9bb