Move ChangeOpStatement away from requiring a StatementListHolder
A first step away from StatementListHolder and ByPropertyIdArray. In
this patch I'm refactoring two private methods to not return arrays
of statements any more, but to directly manipulate the original
StatementList object.

Note: There is an actual change in behavior hidden in this patch. A
StatementList can contain multiple statements with the same GUID. The
old code removed only the first of these. This new implementation will
remove all, but only create an "updateSummary" for the first. However,
this should be purely theoretical. In production it's impossible to
create multiple statements with the same GUID (that's why it's a GUID,
the user can not set it when he creates a new statement).

Bug: T133853
Change-Id: I2421f44349a3ea8b79a7f653c180452b2966839f