Replace meaningless <a class="plainlinks"> with extiw in RC/watchlist
When client links to repo in recent changes and watchlists these links
are currently marked as <a class="plainlinks">. But this class never
had an effect when applied directly to <a>, see
core/resources/src/mediawiki.legacy/shared.css.

As far as I know the intention was to remove clutter (hundreds of
external link icons) from rc/watchlist. But: This can't happen here
because these icons are only applied to external links in wikitext via
<a class="external">. And it never happend, otherwise we would have
seen that.

Replacing with class="extiw" which is for links to sister wikis
(described as "Interwiki Styling" in
core/resources/src/mediawiki.skinning/elements.css).

Discussed and suggested at
https://de.wikipedia.org/wiki/MediaWiki_Diskussion:Common.css

Reviewers: Please test with all skins.

Bug: T48329
Change-Id: Ia004d9cf9f1215e7afbe41771491c0f483c7f63a
Fixes: T52893