Update patch set 2
Patch Set 2:

> Grouping statements is something Repo should provide to View, not the other way round.

This does not change. What changes is the format. Before: Repo provides a "dispatching grouper" that contains the knowledge about groups. After: Repo uses a tool (provided by view) to create a "grouper factory" that contains the knowledge.

View knows that groups exist, but does not know how they are configured. This is the same before and after this patch.

Patch-set: 2