Remove unused Change/ChangeRow/DiffChange methods
The only usage of isEmpty was in ChangeDispatcher::isRelevantChange,
but pointless. A change was "empty" when it did not contained a diff
or an empty diff (which also means that the declaration of isEmpty was
at the wrong place because the idea of a "diff in a change" does not
exist in the Change interface and the ChangeRow class). Anyway. The
later checks in ChangeDispatcher make the previous isEmpty check
pointless, I believe.

Change-Id: I14fcc37d61bd34da9c61f29e6dc23a369e4d85bd