Remove importProperties.php and example CSV files from 2012
The importProperties.php maintenance script is quite obscure: It's a lazy
copy paste from an other script that imported links. It still talks about
"links" everywhere (variables, documentation). It's not doing any "import"
but instead creates new properties with labels it reads from a 1-column CSV
file. The property type is fixed to "wikibase-item". I believe the same can
be done with the special page or other scripts. Fixing this is not worth
our time in my opinion.

simple-elements.csv was introduced in
https://phabricator.wikimedia.org/rEWBA8ac6503ebcb8f68f198516dd04796cfbeba24bff
for use in importInterlang.php which does not exist any more.

en-elements-properties.csv was introduced in
https://gerrit.wikimedia.org/r/27720 for use in importProperties.php.

Change-Id: Iba9db958d0034f7fbb2d7936c96b00ea705da014