Generate autoComments for undo and restore operations
Now there is no longer prefilled summary in the interface where
a user can undo/restore a revision. The field is blank and the
user can add the reason for undoing/restoring the revision which
is then appended to the auto-generated summary.

On the other hand, we can no longer use the native message
"undo-summary". Instead, a new message with quite a similar
content was introduced.

Bug: T41134
Change-Id: I64606be80154e2eb7a924c66d0fba6ccf4c34bda