Fix saving references
This has been broken since I5e85cbf3ebad3dcb670589cefbd13e0267529d3b because
it switched usage from this._lia to this._listview in isValid(), and destroy()
does not clean up this._lia, so isValid() continued to work after destruction.
Instead of reverting the change, I'm preventing isValid() to be called on the
destroyed snaklistview instance.

This should fix the references browser tests.

Change-Id: Idce4ca04351f3f96a1307835cb370e935fc0b7f0