Make ChangeOpStatement work with StatementListProvider w/o index param
As discussed in T133853 the index feature is not used in our code base.
This patch makes the ChangeOp work with MediaInfo as long as there is no
custom index provided.

This does not block T134882. We can still track usage of the index
feature. At the moment it can not be used on non-MediaInfo entity types
anyway.

Bug: T134884
Change-Id: I37563c102dc5f94f73e6583ad6b446f8b46fa368