Move from content model to entity type for namespaces
This changes the way the repoNamespaces and entityNamespaces settings should be
set, but supports old configurations.

DEPLOY: Change repoNamespaces and entityNamespaces for both repo and client
to use entity types instead of content-model ids. Keeping the old style config
will flood the logs with warnings. We should also verify that data access
from the client still functions after this change.

Change-Id: I31b90d2dbec940d60911826725349cdbf2f701ea