Update patch set 1
Patch Set 1: Code-Review+1

As far as I understand start/endAtomic would also work, but could lead to the situation where the actual delete will be delayed because it is stuck in a single transaction with thousands of usage updates. DeferredUpdates avoids that.

Are these Wikibase usages even visible to a user? I don't think so. So it would not matter match if updating the Wikidata usages is delayed. But it's visible to users if the actual delete is delayed.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1