Update patch set 4
Patch Set 4: Code-Review+1

If nobody objects I suggest to merge this within the next few days. Having or not having this will become critical. But having the new codes at the perfect right position is not so critical. If a position turns out to be wrong we can easily change it later.

Patch-set: 4