Update patch set 11
Patch Set 11:

Adrian, this is a pretty big patch, but I'm missing an explanation of the reasoning behind this. Not that there is something wrong (sound like a good thing to do) but is there something that makes this necessary? A Phabricator ticket? I would like to know this before investing time into an actual review.

Patch-set: 11
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0