Automatic defaults for entityNamespaces and repoNamespaces.
If the local wiki is a repo as well as a client, derive the client
settings entityNamespaces and repoNamespaces from the repository
settings.

This is motivated by repeated confusion while installing test systems.
The idea is that when enabling client functionality on a repo wiki,
no additional manual configuration should be needed, since all
necessary information is already present in the repo settings.

Change-Id: Ib8b111eefe3c595e80d91b3f6267ff30d42fda10