Make terms table compatible with MediaInfo
There are multiple intermediate hacks in this patch. The problem is: When I
split this patch into two (one for storing terms, one for retrieving), it's
still unusable. When you try to edit a label or description, both storing and
retrieving must work.

I think all hacks in PS2 are acceptable and this patch can be merged as it is.

A proper solution for retrieving MediaInfo terms is proposed in I9b54124.

Bug: T134735
Change-Id: I0f7e2e352f5bd33666532552f9c800b1f13ab2a1