Allow deserialization of new EntityIdValue serializations
The serialization is not changed yet. This is still discussed in
https://github.com/wmde/WikibaseDataModel/pull/669 . With this in
place we can change the serializaion any time without breaking
anything.

This is the most "lazy" patch I could think of. Do you think these few
lines are worth to be wrapped in a EntityIdValueDeserializer class?

Note that the legacy way must be supported forever, because our
database may always contain legacy serializations.

Bug: T132592
Change-Id: Ieccecd6126a47c0d4c2370162e58a0aee1f5c10e