Update patch set 2
Patch Set 2: Code-Review-2

Please link this to a ticket. There should be a test for this somewhere. Makes me wonder why there was no test that breaks when this is changed?

Setting a temporary -2 because I believe this must be announced carefully. +2 otherwise, because I also believe this is the right thing to do.

Patch-set: 2