Introduce EntityIdComposer and "entity-id-composer"
This is meant as a clean implementation for the hack you can see in
https://gerrit.wikimedia.org/r/#/c/290461/1/lib/includes/Store/Sql/SqlEntityInfoBuilder.php

This goes along with https://github.com/wmde/WikibaseDataModel/pull/670
but is not an exact mirror of the Int32EntityId interface proposed there.
This builder is much more general and also allows constructing all kinds of
entity IDs that contain fixed parts that never change for that entity type
(e.g. "File:Filename.jpg" where "File:" is a fixed prefix, but the suffix
is not a number).

The new stuff is not yet used in this patch. It's meant to replace all
usages of LegacyIdInterpreter::newIdFromTypeAndNumber as well as remaining
…::newFromNumber when it makes sense.

Bug: T136294
Change-Id: I9b541240ccbfb22425ab48b7de9c635be14dfc6d