Fix display of outdated label and description in EntityView header
Problem was use of EntityInfoTermLookup, which retrieves terms from
the terms table (via SqlEntityInfoBuilder).

When an edit is made, ParserOutput (including page html) is generated
before save and before the terms table is updated. So, whatever the
terms were in the previous revision of an entity is what was displayed.

This changes the code to use an EntityRetrievingTermLookup, with an
InMemoryEntityLookup (populated with the entity that is being saved).
This allows the language fallback to still be applied.

More ideal would be a solution with less layers of indirection,
and maybe not use LabelDescriptionLookup, though I think more refactoring
is needed to make this possible.

This is the smallest patch possible to fix the bug.

I am not sure how best to add tests for this, other than to have
a browser test that covers this and maybe we already have one?

DEPLOY: there is an off-chance that this may affect the operation of
SpamBlacklist and AbuseFilter. We did our best to check for this, but
we should send a heads-up to wiki admins when deploying this, and we
should double-check that the filters work after deployment.

Bug: T135714
Change-Id: I1062d4451eca0218b3481fe2c13743273552fb4d
(cherry picked from commit ed1ca9bbc886e1bf00a807fea9cc2c5272ce7c7f)