Update patch set 4
Patch Set 4:

NOTE: Aude pointed out that we need to ensure compatibility with SpamBlacklist and AbuseFilter. Such extensions may force Entities to be rendered before saving, so they can apply filters. This may mean rendering an Entity before an EntityId is known. To resolve this, we could produce a ParserOutput without HTML: filters are typically not applied to the actual HTML, but other pars of ParserOutput, such as the list of external links.

Patch-set: 4