Update patch set 5
Patch Set 5:

Ignore above comment, it was intended for I5a630ca2a5a2.

We need to provide a way for extensions that provide more 
entity types to handle the "no such entity" case differently. 
Idea: have a method in EntityView that handles the "no such entity" case, and override it for MediaInfo.

Patch-set: 5