Update patch set 4
Patch Set 4:

Well, I'm in general wary about hooks. I think about them as a top-level violation of a generalized open-closed principle. It somewhat makes sense for all-embracing frameworks (which are bad) like MediaWiki, but it seems like something to avoid for applications or libraries. WikibaseRepo could in theory be configured top-to-bottom with no need to hook in anywhere.

Patch-set: 4