Update patch set 8
Patch Set 8:

@adrian wrote:
> My understanding is that this is regarded by you as a quick, non-optimal implementation. I'd like to know what your mid-term goal for this code area is.

Refactor all the API modules :)  Honestly, I'm not sure yet. I'll have to expermiment with moving the "fresh id" logic around a bit.

> I would like to have the EntityId handling removed from EntityStore, specifically `assignFreshId` and `canCreateWithCustomId`. I think it would make sense to simplify `saveEntity` so that it always expects entities with id.

I agree that saveEntity should simply require an ID to be already set. But the calling code then needs to know if and how an ID can be assigned when creating a new entity. They could use the EntityHandler for that, but that creates a strong binding to mediawiki in general and wiki pages in particular. I would like to avoid that. So assignDFreshId and supportsIncrementalId and canCreateWithCustomId that exist in the EntityHandler would need to be repeated in some other service interface (this is basically framework For now, I have put this into the EntityStore, but I would like to revisit that decision.

> I looked at Api/EditEntity and Api/ModifyEntity. They already create new entities (via entityFactory) and call assignFreshId on them. That could call through to an entity type specific function (in contenthandler?) and pass the serialized entity id string from the serialized entity blob.

Yes, I agree. That's not the tricky part. The tricky part is to know when an automatic id can be used, and where to get it from. And also, what custom IDs are OK for creation.

Patch-set: 8