Update patch set 2
Patch Set 2:

I think that's not what I meant. I meant that ModifyEntity doesn't even know about whether something gets assigned a fresh id or anything. It just passes the serialized entity id from the JSON (if present) and whether `new` was set to a service that either throws or returns an EntityId. That service can then a) Decide if the entity id given is valid and a new entity can be created on the fly or b) a fresh entity id has to be allocated.

Patch-set: 2