Update patch set 3
Patch Set 3:

looks good to me, yes! If we release 6.1 now, it will automatically get used, right? Will that break the tests? do we need to pin wikibase to using 6.0 first, then do the release, and then update this patch here to requiring 6.1, and adapting the tests?

Patch-set: 3
Label: Code-Review=0