Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

All three comments are meant as TODOs. I suggest to continue working on this code in later patches, but leave this as small as it is. No +2 to give Daniel some time to see this. I will merge this soon if nobody objects.

Patch-set: 3
Label: Code-Review=+1