Update patch set 2
Patch Set 2: Code-Review+1

Code is ok, but this still needs a PM decision and an announcement: One JS config variable is replaced with an other (with a slightly different meaning). And it's not possible any more to have a Babel box with a single language. ULS will always add other languages, up to 4. With this change this behavior can only be disabled with a Babel box with 2 or more languages.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1