Update patch set 9
Patch Set 9:

I uploaded a follow-up with all issues fixed, except one: I believe the name of the canCreateWithCustomId method is not correct. From what I see it does not check if an entity exists or not. It will always return true if an entity can be edited, no matter if it exists or not. Therefor I suggest to rename it *or* to clarify the contract/docs of that method.

Patch-set: 9