Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

CSS is fine for me now. Thanks. I also love the idea of longer tooltips (e.g. "remove this statement with it's qualifiers and references"). But this should be done in an other patch, obviously. Is there an easy way to stick to the current "remove" tooltip, as it was before this patch?

Patch-set: 3
Label: Code-Review=+1