Fix siteselector not closing after selecting a site
This is a direct follow up to I575f1e6. It introduced a bug: In some
selectors (e.g. the siteselector, where you select the site for a sitelink)
the input field is automatically refocused after a suggested item was
selected. This reopened the selector, causing flickering and confusion.

With this patch the selector is only reopened when you use the mouse, but
not via tab key. I believe this is ok: when you use the keyboard, you can
press "cursor down" to reopen the selector. But it's impossible with the
mouse alone.

Bug: T135008
Change-Id: Iad68c459d56554f47e6c34c377499297e3d37e66