Add min-width for site blocks in sitelinklistview
This fixes a minor display issue in the sitelinks section that becomes
visible when there are very short site names (e.g. "it") together with
longer ones (typically "en").

This replaces
https://www.wikidata.org/wiki/Special:Diff/190972736

I'm using 1.5em because 1.4em is not enough to include "en" and "de".

I'm setting this on the inner <div> so it is active in both read and
edit more.

Change-Id: I5e4173376990bb38cb2b4df28fa086d9cbcf8529