Do not show calendar model with precision 10 years or larger
The calendar model *is* critical with precision month, day and so on.
It *may* be relevant with precision year.
In both cases we enforce displaying the calendar model when it's not
the same as the one our parser guesses.

But is is *not* relevant with precision 10 years, 100 years and so on.
It just does not make any difference.

Bug: T133973
Change-Id: Ic891718364c6a6c53beb48fa10adae5ad0f7a7fa