Update patch set 1
Patch Set 1:

Correct, but most probably not relevant. One of the first calendar models we will implement is the Julian day (number of days since 4713 BC). This will most probably not support precisions other than day. Besides, other calendar models that are not compatible with the Gregorian idea of a "year" must *always* show the calendar model indicator. This is already supported by this class.

Patch-set: 1