Update patch set 5
Patch Set 5:

This is a very weird test. I'm not sure if this should be in our code base as it is now. Will it help us or cause more trouble when it starts breaking and people don't exactly what the test is meant to do and know how to fix it?

Patch-set: 5