Update patch set 4
Patch Set 4: Code-Review+1

This is a great idea and the code looks much better with this.
We still need to test this in detail, because in my experience our counters and edit states can be a mess.

Patch-set: 4
Reviewer: Jonas Kress (WMDE) <2677@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1