Update patch set 14
Patch Set 14:

(1 comment)

The architecture looks good now, but looking at the actual RFC output in QuantityRdfBuilderTest.php, I'm no longer sure this is the right way to go.

As this patch stands, it generates something like
  wd:Q1 psv:P7 wdv:123abc;
  wdv:123abc a wikibase:QuantityValue
    wikibase:quantityAmount "+1211"^^xsd:decimal ;
    wikibase:quantityUnit   "cm" .
  wdv:123abc wikibase:quantityNormalized wdv:345cde .
  wdv:wdv:345cde a wikibase:QuantityValue
    wikibase:quantityAmount "+12"^^xsd:decimal ;
    wikibase:quantityUnit   "m" .

So the normalized value is chained to the actual value. I would rather like to have them on the same level:

  wd:Q1 psv:P7 wdv:123abc;
  wd:Q1 psvn:P7 wdv:345cde;
  wdv:123abc a wikibase:QuantityValue
    wikibase:quantityAmount "+1211"^^xsd:decimal ;
    wikibase:quantityUnit   "cm" .
  wdv:wdv:345cde a wikibase:QuantityValue
    wikibase:quantityAmount "+12"^^xsd:decimal ;
    wikibase:quantityUnit   "m" .

The prefix psvn would work just like psv: if psv stands for <http://www.wikidata.org/prop/statement/value/>,
psvn stands for <http://www.wikidata.org/prop/statement/value-normalized/>. So we will define an extra repo-specific prefix instead of introducing a new vocab term.

The modeling you propose is a bit more flexible, but more annoying to use in queries. Since our primary use case for having this is querying, it would perhaps be nice to make it as easy as possible to write queries against normalized values. With the second proposal, it's not necessary to ever tough the un-normalized value if you don't want to. In cases where no normalization was applied, psv and pswv would just point to the same resource.

Patch-set: 14