Minimize EntityIdValue footprint when storing to the database
Issue with this proposal: The knowledge about these array keys is
disjoint and repeated in two different components. The plan is to
have a breaking change in the EntityIdValue class one day that drops
the legacy array keys everywhere. The same moment the class introduced
in this patch becames obsolete.

Bug: T56085
Change-Id: Iec661499b46b21651625c56a978c6eabd172bd26