Update patch set 10
Patch Set 10:

I changed this back to connecting to T140760. This *is* needed for on-demand creation when editing a label. I was confused about this because there is another incomplete patch that implements on-demand creation for statement edits (Id219eca29b0), but that is still a draft, and thus invisible to most.

Patch-set: 10
Label: Verified=0