Update patch set 3
Patch Set 3:

this is definitely an improvement and suppose okay to merge, but when i search for a value and none is found then it says "No match was found" but then also still has the "more" link.

Can we not show the more link when there are no matching search results?

Patch-set: 3
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0