Update patch set 2
Patch Set 2:

I suggest to go the route I outlined in T115269#2511555.

It's true that we have a QuantityValue class in JavaScript, and we should adopt it. But as far as I can tell it is currently unused. The expert does not know anything about the string the user enters. It passes the input to the parser API, passes the resulting JSON to the formatter API and shows the resulting string. That's all.

Patch-set: 2
Label: Verified=0