Update patch set 5
Patch Set 5:

I agree with Thiemo: the generic SPARQL client code should not be in the Wikibase component. It should probably be in a separate component, or tied together with API client code.

In the long run, I'm not convinced we should rely on conversion factors provided by wikidata items. That's what this is doing, right? 

I'd much rather see us use GnuUnits to generate our config. We'll need a mapping from Q-IDs to "canonical" unit symbols for that, though.

Patch-set: 5
Label: Code-Review=0