Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

As pointed out by @hoo on IRC: we actually need to be able to specify the fallback mode at the same time as a specific revision. We can't get away with using the same parameter for the revision ID and the fallback flag, we need to have separate parameters for those.

Patch-set: 1
Label: Code-Review=-1