Update patch set 14
Patch Set 14: Code-Review-1

(20 comments)

Dynamically building an SQL query in a way that the code is well readable is really hard, I can see that. What you did so far is really good. Thanks a lot! I would love to help you getting this in a state that fits the project's standards. I can, for example, continue working on this and upload an other patch set, if this is ok for you and you want me to do that. Please tell me.

Patch-set: 14
Label: Code-Review=-1