Make EntitySavingHelper extend EntityLoadingHelper.
Rationale: The idea is that the logic for on-the-fly auto-creation of entities should be
encapsulated in the EntitySavingHelper. However, with the current code structure,
this needs soem kind of intervention during loading as well as during saving.
To keep this knowledge local, loading and saving must be handled by the same class.
It may even become necessary to maintain state in the helper (e.g. the base revision
id to avoid conflicts).

Justification: Subclassing seems ok here, since saving makes no sense without loading.
Per the above argument, loading and saving should be done via the same object, so the
saving helper actually IS A loading helper.

Also, EntitySavingHelper and EntityLoadingHelper come from the same factory, which
knows all relevant services to be injected.

Bug: T140760
Change-Id: I3a019f558effd411bef165fbcbdc824aa958b447