Update patch set 1
Patch Set 1: Code-Review+1

I'm not sure if it's a good idea to make the two tests based on each other. I learned this tends to cause more harm than good. But appart from that this patch is absolutely lovely. I did not expected this refactoring to be that obvious. Cool.

Patch-set: 1