Remove bogus CSS selectors when statementview shows an error
This is a direct follow up to I706dbe8. Steps to reproduce: Add a new
statement (with the last "add" button). Pick a monolingual property (or
an other that allows to enter invalid stuff). Enter an invalid language.
The statementview will turn red, but there is still a blue box behind
the (now read-only) property label.

Bug: T141862
Change-Id: Ie3bf8f646581840a57e45bbda0910855a0adcd9a