Skip property label scrolling on statement groups with one statement
This is a direct follow up to Ic52065d.

This removes all bogus logging code (note that the number was always 0).

This replaces the possibly expensive ….is( ':empty' ) check with a super
trivial ….length check. This does have the following effect:
findFirstVisibleMainSnakElement() is called with a statementlistview (a
set of statements inside a statementgroupview, guaranteed to have the
same property). If such a list contains a single main snak only, there is
nothing to scroll. We can return null and jQuery.add() will ignore null.

Typically there are many, many statement groups with only one statement.

This includes the new, empty statement that appears when using the "add"
function.

Change-Id: I94b9173d6c12f450a93c3444935c49ae1cf97b15