Don't abuse ParserOptions::getUser()->getOption( 'language' ) to get the user lang
As that's not necessary the desired user language, as that can
also be altered via ?uselang. This is especially relevant for
logged-out users.

I removed the hack in question by simply using ParserOptions::getUserLangObj.
In order to not split the parser cache by user language where not
needed, I divided WikibaseLuaBindings into one class with language
dependent functionality and one with language independent functionality,
so that these can be instantiated independently.

Please note, that the hack in question wasn't needed in Scribunto_LuaWikibaseEntityLibrary
in the first place, as the only way to obtain a mw.wikibase.entity
object is to call mw.wikibase.getEntity/getEntityObject which already
splits the parser cache if "allowDataAccessInUserLanguage" is true.

This also adds tests for WikibaseLanguageIndependentLuaBindings::getSetting.

Bug: T142906
Change-Id: Iecc1fa18124d655bbce554d5ce7b33b67ee1abe7