Update patch set 7
Patch Set 7: Code-Review-1

The proposed scheme introduces a new entity "wdata:Q2 a schema:Dataset" that is connected to the relevant entity by one level of indirection. It would be better for performance and usability if the properties instead would be added directly to the entity, since this will lead to shorter, simpler queries. There is no risk of confusion since the page properties are clearly distinguished from the entity data properties.

Patch-set: 7
Label: Code-Review=-1