Set <meta description> on entity pages
This is a proof of concept.

I may miss one of the consequences this haves. Is this a possible attack
vector? Do the descriptions we have really fit well into this meta tag?
Will this make the search results in external search engines always better,
or worse in other situations?

Personally I think this is fine. The only modification I can think of is
a static prefix or suffix in this <meta description> field, e.g. something
like "<description goes here> - facts and sources".

Bug: T88475
Change-Id: Ic3ca34a6d0f57431249ce4cda6a18b941e3ccf31