Update patch set 30
Patch Set 30: -Code-Review

I verified manually that this behaves as expected, and the tests in the follow-up look good.

However, looking at the data returned by this module, the structure seems to be rather inconvenient:

  "wbentityusage": [
                    {
                        "aspect": "O",
                        "*": "Q7"
                    },
                    {
                        "aspect": "S",
                        "*": "Q9"
                    }
                ]

I suppose what the caller is usually most interested in is a list of item IDs. This is in there, but hard to come by. I would suggest something like this:

  "wbentityusage": [
                    "Q7":["O"],
                    "Q9":["S"]
                ]
                
This seems a lot more streight forward. A more future-proof compromize would perhps be:

  "wbentityusage": [
                    "Q7":{ "aspects":["O"] },
                    "Q9":{ "aspects":["S"] }
                ]

What do you think?

Patch-set: 30
Label: Code-Review=0