Update patch set 6
Patch Set 6:

We should be carefully about adding more tests to run on every commit to not extend the waiting time for the jobs too much. Of course we want to add more tests to run regularly on beta or test as we are adding new features.

Patch-set: 6
Reviewer: Tobias Gritschacher <170@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0