Use one atomic section per function, not per batch in EntityUsageTable
It makes little sense to use one atomic section for each batch,
rather than one atomic section for the whole operation. If we
hit an error, we want to whole operation to be rollbacked, not
just the current batch.

Should this cause problems with long lasting transactions,
the atomic sections should be removed instead.

Change-Id: I25eb9a8509377938a1d88ef56128da3b067551b3