Enable caching per default for Api modules with static results
All of these should return values that hardly ever changes, thus
it should be safe to cache them for an hour.

This could also be done for FormatSnakValue/ParseValue when
the results don't include any entity labels. We could cache
results depending on the data type there, but I'm not sure
it's worth it.

Change-Id: I722969076fbae6a0c0969b89358d4e2ee53e940f