More specific #toc selector
I just saw this empty rectangle on the live system, while using
debug=1. This happened because the CSS selector is not more specific
than the original one. Which one is active depends on the loading
order of the files. With a more specific selector the loading order
does not matter any more.

Change-Id: I7151f166c24a81f0c21035b2b8e3e1a313c82e5c