Extract subdomain when pasting full URL into siteselector
(Alternative to I62578bdfac821ec1030dde4bfca4fbf25c488da8, see below)

This is a trivial first step towards the idea to allow pasting full
Wikipedia URLs into the sitelinks editor. This patch does have a
single effect: Before, nothing was found when a user tried to paste a
URL. Now, it finds the site (but not the page).

This has been adopted from I62578bdfac821ec1030dde4bfca4fbf25c488da8
by Marius Hoch, so that it matches the pasted URL with the site URL,
instead of trying to match the pasted URL with the site id.

Bug: T144310
Change-Id: Ifb5b3b2f2cbef186f7349c5c65964ee430fe626d