Update patch set 2
Patch Set 2:

The \w already includes underscores. Which does not matter much because our subdomains don't use underscores. So there is nothing I can do at the moment to improve this. It works fine in 99% of the cases and can be improved later.

The only issue is that this misses tests.

Patch-set: 2
Label: Verified=0