Update patch set 8
Patch Set 8: Code-Review+1

(1 comment)

I can confirm, but personally I find this even better than the old behavior. When I remove the last snak from a reference, the reference header with it's remove button and the "add" button to add a snak to the reference is still there. I find this very nice and helpful.

This empty reference goes away (as it should be) when I save.

With the old behavior the whole reference header was removed. This caused a big visual change. Stuff I did not clicked and did not meant to remove disappeared. This was potentially confusing.

Patch-set: 8
Label: Code-Review=+1