Update patch set 8
Patch Set 8:

Please note that this is the wrong place to discuss the architecture of the DataValues Number component. This was mostly done in https://github.com/DataValues/Number/pull/66 and related pull requests. Please open a Phabricator ticket if you think this needs a redesign and an other breaking change (could actually be a good argument for a DataValues Number 1.0 release).

I don't see the "Square extends Rectangle" problem here because nothing the subclass does changes the meaning and behavior of the base class.

The fact that this is about code sharing is even explicitly highlighted in the documentation of these classes. Does it help if we mark them with "final" when possible?

I agree on the naming. A "QuantityValue" base class with no bounds would be a bit nicer. But as Daniel explained this would be a major breaking change and we decided against this.

Patch-set: 8
Label: Verified=0