Update patch set 2
Patch Set 2:

The code touched in this patch is, in essence, a multi-step "garbage in, entity ID out" parser. There are two levels of knowledge encoded in this code:
* Knowledge where to look for an entity ID in the input string.
* Knowledge how an entity ID might look like.

Level 1 can not be somewhere else. The input string can be anything and I really want a parser for anything, not only for specific formats like URIs.

Level 2:
* Entity IDs are made of ASCII word characters.
* Entity IDs are at least 2 characters long.

Both assumptions are currently true, and I believe will always be true in most, if not all cases. And even if not, the code will continue to work as before and allow to search for entity IDs. This code path is untouched.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0