Skip UnitConverterTest cases if bcmath is not available
This makes exact matches on expected products, which DecimalMath
can only give us if it has bcmath.

Given we use DecimalMath/ strings to internally represent these values,
we could make DecimalMath better here, but I'm not sure that's worth
the hassle. I played with that, but I still couldn't get it right for
all test cases here (but for most).

Change-Id: I6dc78ac5035b623d25e35f4eb9fb965950e4a272