Update patch set 5
Patch Set 5:

The problem with GNU units config is that it's pretty hard to get from it to the units in Wikidata, and that's what we actually need. I.e. how we know what GNU Units calls Q26250779 aka "square arshin"?

Anyway, nobody actually forces us to be married to this script. We can use it to bootstrap the data and if we don't like how it works we can use another one later.

Patch-set: 5