Update patch set 5
Patch Set 5:

About SPARQL client I'm kind of ambivalent. It is a super-basic client (I mean, tiniest wrapper over Http::get) so if we position it as generic one we'll need it to do more. It's good enough for units script, but if we want to do more sparql we'd need something stronger. OTOH, I don't want to make a full-blown lib out of it because there already are a bunch of sparql clients out there and maintaining a separate library just for couple of functions is really an overkill. 

So, would be putting it into includes/ good enough?

Patch-set: 5
Label: Verified=0